API Remove deprecated code
