+++ _config/asset.yml
@@ -29,7 +29,6 @@ SilverStripe\Core\Injector\Injector:
       ResolutionFileIDHelpers:
         - '%$SilverStripe\Assets\FilenameParsing\HashFileIDHelper'
         - '%$SilverStripe\Assets\FilenameParsing\NaturalFileIDHelper'
-        - '%$SilverStripe\Assets\FilenameParsing\LegacyFileIDHelper'
       DefaultFileIDHelper: '%$SilverStripe\Assets\FilenameParsing\NaturalFileIDHelper'
       VersionedStage: Live
   # Define protected resolution strategy
