+++ src/FilenameParsing/NaturalFileIDHelper.php
@@ -7,8 +7,7 @@ use SilverStripe\Core\Injector\Injectable;
 /**
  * Parsed Natural path URLs. Natural path is the same hashless path that appears in the CMS.
  *
- * Natural paths are used by the public adapter from SilverStripe 4.4 and on the protected adapter when
- * `legacy_filenames` is enabled.
+ * Natural paths are used by the public adapter from SilverStripe 4.4
  *
  * e.g.: `Uploads/sam__ResizedImageWzYwLDgwXQ.jpg`
  */
