--- tests/php/Dev/Tasks/FileMigrationHelperTest.yml
@@ -1,71 +1,0 @@
-SilverStripe\Assets\Folder:
-  parent:
-    Name: ParentFolder
-  subfolder:
-    Name: SubFolder
-    Parent: =>SilverStripe\Assets\Folder.parent
-  uploads-folder:
-    Name: Uploads
-SilverStripe\Assets\Image:
-  image1:
-    Name: myimage.jpg
-  image2:
-    Name: myimage.jpg
-    ParentID: =>SilverStripe\Assets\Folder.subfolder
-  ## SS2 would sometime treat PDF files as images
-  fakePdf:
-    Name: myimage.pdf
-  ## Image with an invalid SS4 name
-  badimage:
-    Name: bad__image.gif
-SilverStripe\Assets\File:
-  file1:
-    Name: anotherfile.txt
-  file2:
-    Name: file.doc
-    ParentID: =>SilverStripe\Assets\Folder.parent
-  file3:
-    Name: picture.pdf
-    ParentID: =>SilverStripe\Assets\Folder.subfolder
-  # The front end will want to rename this file
-  multi-dash-file:
-    Name: multi-dash--file---0.pdf
-    ParentID: =>SilverStripe\Assets\Folder.subfolder
-  # We won't create a physical file for entry. Our process should carry on.
-  missing-file:
-    Name: missing-file.pdf
-  # SS4 doesn't allow files with double underscores
-  badname:
-    Name: bad__name.zip
-    ParentID: =>SilverStripe\Assets\Folder.parent
-  badname2:
-    Name: bad__0.zip
-  # EXE files are not allowed
-  invalid:
-    Name: myfile.exe
-    ParentID: =>SilverStripe\Assets\Folder.subfolder
-  # Files with no extension are not allowed
-  no-extension:
-    Name: myfile
-    ParentID: =>SilverStripe\Assets\Folder.subfolder
-  # badnameconflict will want to override goodnameconflict
-  goodnameconflict:
-    Name: bad_name.doc
-  badnameconflict:
-    Name: bad__name.doc
-  # name with wrong case - these file will be written to disk with a different case fileID
-  wrongcase:
-    Name: wrong-case.txt
-  mismatch-folder-case:
-    Name: good-case-bad-folder.txt
-    ParentID: =>SilverStripe\Assets\Folder.uploads-folder
-  good-match-folder-case:
-    Name: good-case-good-folder.txt
-    ParentID: =>SilverStripe\Assets\Folder.uploads-folder
-  too-many-case:
-    Name: too-many-alternative-case.txt
-  # Case conflict, 2 files with the same name in different case
-  all-uppercase:
-    Name: MIXED-CASE-FILE.TXT
-  all-lowercase:
-    Name: mixed-case-file.txt
