--- tests/php/Dev/Tasks/FileMigrationHelperTest/Extension.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace SilverStripe\Assets\Tests\Dev\Tasks\FileMigrationHelperTest;
-
-use SilverStripe\Assets\File;
-use SilverStripe\Dev\TestOnly;
-use SilverStripe\ORM\DataExtension;
-
-/**
- * @property File $owner
- */
-class Extension extends DataExtension implements TestOnly
-{
-    /**
-     * Ensure that File dataobject has the legacy "Filename" field
-     */
-    private static $db = [
-        "Filename" => "Text",
-    ];
-
-    public function onBeforeWrite()
-    {
-        // Ensure underlying filename field is written to the database
-        $this->owner->setField('Filename', 'assets/' . $this->owner->generateFilename());
-    }
-}
