--- tests/php/Dev/Tasks/FixFolderPermissionsHelperTest.yml
@@ -1,11 +1,0 @@
-SilverStripe\Assets\Folder:
-  parent:
-    Name: ParentFolder
-    CanViewType: NULL
-  subfolder:
-    Name: SubFolder
-    Parent: =>SilverStripe\Assets\Folder.parent
-    CanViewType: Anyone
-  anotherfolder:
-    Name: AnotherFolder
-    CanViewType:
