--- tests/php/Dev/Tasks/LegacyThumbnailMigrationHelperTest.yml
@@ -1,19 +1,0 @@
-SilverStripe\Assets\Folder:
-  parent:
-    Name: ParentFolder
-  subfolder:
-    Name: SubFolder
-    Parent: =>SilverStripe\Assets\Folder.parent
-SilverStripe\Assets\Image:
-  toplevel:
-    Name: toplevel.jpg
-  nested:
-    Name: nested.jpg
-    ParentID: =>SilverStripe\Assets\Folder.subfolder
-  nested-sibling:
-    Name: nested-sibling.jpg
-    ParentID: =>SilverStripe\Assets\Folder.subfolder
-SilverStripe\Assets\File:
-  file1:
-    Name: anotherfile.jpg
-
