--- tests/php/Dev/Tasks/NormaliseAccessMigrationHelperTest.yml
@@ -1,22 +1,0 @@
-SilverStripe\Security\Group:
-  file-master:
-    Title: File Master
-
-SilverStripe\Assets\Folder:
-  parent:
-    Name: Uploads
-  restricted:
-    Name: Restricted
-    FileFilename: Restricted
-    CanViewType: OnlyTheseUsers
-    ViewerGroups:
-      - =>SilverStripe\Security\Group.file-master
-SilverStripe\Assets\File:
-  file1:
-    FileFilename: file1.txt
-    Name: file1.txt
-  secret:
-    FileFilename: Restricted/secret.txt
-    Name: file1.txt
-    CanViewType: Inherit
-    Parent: =>SilverStripe\Assets\Folder.restricted
