--- tests/php/Dev/Tasks/SecureAssetsMigrationHelperTest.yml
@@ -1,23 +1,0 @@
-SilverStripe\Assets\Folder:
-  parent-with-protected-sub:
-    Name: parent-with-protected-sub
-  protected-sub:
-    Name: protected-sub
-    CanViewType: LoggedInUsers
-    Parent: =>SilverStripe\Assets\Folder.parent-with-protected-sub
-  unprotected:
-    Name: unprotected
-  protected:
-    Name: protected
-    CanViewType: LoggedInUsers
-  protected-inherited:
-    Name: protected-inherited
-    CanViewType: Inherit
-    Parent: =>SilverStripe\Assets\Folder.protected
-  protected-with-modified-htaccess:
-    Name: protected-with-modified-htaccess
-    CanViewType: LoggedInUsers
-SilverStripe\Assets\Image:
-  nested:
-    Name: nested.jpg
-    ParentID: =>SilverStripe\Assets\Folder.protected-sub
