--- tests/php/Dev/Tasks/TagsToShortcodeHelperTest.yml
@@ -1,48 +1,0 @@
-SilverStripe\Assets\Tests\Dev\Tasks\Shortcode\PseudoPage:
-  page1:
-    Title: 'my page'
-    Content: >
-      <p id="filelink">file link <a href="/assets/document.pdf">link to file</a></p>
-      <p id="image"><img src="/assets/myimage.jpg"></p>
-      <p id="variant"><img src="/assets/_resampled/ResizedImageWzY0LDY0XQ/myimage.jpg" width="64" height="64"></p>
-
-SilverStripe\Assets\Folder:
-  decadefolder:
-    Name: decade80
-
-SilverStripe\Assets\Image:
-  image1:
-    Name: myimage.jpg
-    FileFilename: myimage.jpg
-    FileHash: 33be1b95cba0358fe54e8b13532162d52f97421c
-  underscoreFile:
-    Name: under_score.jpg
-    FileFilename: decade80/under_score.jpg
-    FileHash: 33be1b95cba0358fe54e8b13532162d52f97421c
-    Parent: =>SilverStripe\Assets\Folder.decadefolder
-  trippleUnderscoreFile:
-    Name: under_score.jpg
-    FileFilename: decade80/under_score-v2.jpg
-    FileHash: 33be1b95cba0358fe54e8b13532162d52f97421c
-    Parent: =>SilverStripe\Assets\Folder.decadefolder
-SilverStripe\Assets\File:
-  document:
-    Name: document.pdf
-    FileFilename: document.pdf
-    FileHash: 0ba2141b8996a615d7484536d7a97c27a1768407
-
-SilverStripe\Assets\Tests\Dev\Tasks\Shortcode\HtmlObject:
-  htmlObject:
-    HtmlLine: '<a href="/assets/document.pdf">link to file</a>'
-    HtmlLineNoShortCode: '<a href="/assets/document.pdf">This wont be converted</a>'
-    Content: '<a href="/assets/document.pdf">Content Field</a>'
-
-SilverStripe\Assets\Tests\Dev\Tasks\Shortcode\SubHtmlObject:
-  subHtmlObject:
-    HtmlLine: '<a href="/assets/document.pdf">HtmlObject Table</a>'
-    HtmlContent: '<img src="/assets/myimage.jpg" alt="SubHtmlObject Table">'
-    Content: '<a href="/assets/document.pdf">Content Field</a>'
-
-SilverStripe\Assets\Tests\Dev\Tasks\Shortcode\NoStage:
-  stageless:
-    Content: '<a href="/assets/document.pdf">Stageless Versioned Object</a>'
