--- tests/php/Storage/AssetStoreTest/TestAssetStore.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace SilverStripe\Assets\Tests\Storage\AssetStoreTest;
-
-use SilverStripe\Assets\Dev;
-
-/**
- * Allows you to mock a backend store in a custom directory beneath assets.
- * @deprecated use SilverStripe\Assets\Dev\TestAssetStore instead
- */
-class TestAssetStore extends Dev\TestAssetStore
-{
-
-}
