+++ tests/php/Storage/DefaultAssetNameGeneratorTest.php
@@ -17,7 +17,7 @@ class DefaultAssetNameGeneratorTest extends SapphireTest
      */
     public function testWithoutPrefix()
     {
-        Config::modify()->merge(DefaultAssetNameGenerator::class, 'version_prefix', '');
+        Config::modify()->set(DefaultAssetNameGenerator::class, 'version_prefix', '');
         $generator = new DefaultAssetNameGenerator('folder/MyFile-001.jpg');
         $suggestions = iterator_to_array($generator);
 
@@ -64,7 +64,7 @@ class DefaultAssetNameGeneratorTest extends SapphireTest
 
     public function testPathsNormalised()
     {
-        Config::modify()->merge(DefaultAssetNameGenerator::class, 'version_prefix', '-v');
+        Config::modify()->set(DefaultAssetNameGenerator::class, 'version_prefix', '-v');
         $generator = new DefaultAssetNameGenerator('/some\folder/MyFile.jpg');
         $suggestions = iterator_to_array($generator);
         $this->assertEquals(100, count($suggestions ?? []));
@@ -80,7 +80,7 @@ class DefaultAssetNameGeneratorTest extends SapphireTest
      */
     public function testWithDefaultPrefix()
     {
-        Config::modify()->merge(DefaultAssetNameGenerator::class, 'version_prefix', '-v');
+        Config::modify()->set(DefaultAssetNameGenerator::class, 'version_prefix', '-v');
 
         // Test with item that doesn't contain the prefix
         $generator = new DefaultAssetNameGenerator('folder/MyFile-001.jpg');
@@ -120,7 +120,7 @@ class DefaultAssetNameGeneratorTest extends SapphireTest
 
     public function testFolderWithoutDefaultPrefix()
     {
-        Config::modify()->merge(DefaultAssetNameGenerator::class, 'version_prefix', '');
+        Config::modify()->set(DefaultAssetNameGenerator::class, 'version_prefix', '');
         $generator = new DefaultAssetNameGenerator('folder/subfolder');
         $suggestions = iterator_to_array($generator);
 
@@ -134,7 +134,7 @@ class DefaultAssetNameGeneratorTest extends SapphireTest
 
     public function testFolderWithDefaultPrefix()
     {
-        Config::modify()->merge(DefaultAssetNameGenerator::class, 'version_prefix', '-v');
+        Config::modify()->set(DefaultAssetNameGenerator::class, 'version_prefix', '-v');
         $generator = new DefaultAssetNameGenerator('folder/subfolder');
         $suggestions = iterator_to_array($generator);
